+++ packages/kit/src/extensions/feature/index.ts
@@ -1,1 +1,3 @@
+export * from "./capability";
 export * from "./FeatureTool";
+export * from "./FeatureCapabilityExtension";